#336 cleanup and documentation
